Move events to separate folder
