+++ packages/game/src/events/eventBuilder.ts
@@ -1,0 +1,4 @@
+import { createEventDefinition } from 'ts-bus'
+
+export const buildEvent = <T>(eventName: string) =>
+  createEventDefinition<T>()(eventName)
