+++ packages/game/src/events/index.ts
@@ -1,0 +1,3 @@
+export * from './eventBuilder'
+export * as clientEvents from './clientEvents'
+export * as ServerEvents from './serverEvents'
