+++ packages/game/src/index.ts
@@ -4,8 +4,7 @@ import * as Components from './components'
 import * as State from './state'
 import * as Types from './types'
 
-export * as ClientEvents from './clientEvents'
-export * as ServerEvents from './serverEvents'
+export * from './events'
 
 export * as Systems from './systems'
